+++ app/javascript/dashboard/components/ui/ContextMenu.vue
@@ -2,7 +2,7 @@
 import { computed, onMounted, nextTick, useTemplateRef } from 'vue';
 import { useWindowSize, useElementBounding } from '@vueuse/core';
 
-import CustomTeleport from 'dashboard/components-next/CustomTeleport.vue';
+import TeleportWithDirection from 'dashboard/components-next/TeleportWithDirection.vue';
 
 const props = defineProps({
   x: { type: Number, default: 0 },
@@ -59,7 +59,7 @@ onMounted(() => {
 </script>
 
 <template>
-  <CustomTeleport to="body">
+  <TeleportWithDirection to="body">
     <div
       ref="menuRef"
       class="fixed outline-none z-[9999] cursor-pointer"
@@ -69,5 +69,5 @@ onMounted(() => {
     >
       <slot />
     </div>
-  </CustomTeleport>
+  </TeleportWithDirection>
 </template>
